Update pagespeed.conf
With Nginx 1.9 and above and pagespeed above 1.11.32, you need to move:

pagespeed  FileCachePath  "/var/tmp/";
pagespeed  LogDir "/var/log/pagespeed";

to nginx.conf so that additional code can be added. Specifically, to be able to add code to allow Chrome iOS to see images as webp is no longer supported.